Specialize ArlasTest to data providing.
ArlasTest now provides only test data,
So utils have been moved to a Helper util.
FlowFragment data are provided by ArlasTest
to be used for further tests.
Data itself
is provided by a data generator class; using
this kind of class will help us to provide
more complicated test data by keeping
the code separated and readable.
