+++ src/test/scala/io/arlas/data/transform/features/WithDurationFromIdTest.scala
@@ -23,12 +23,13 @@ import io.arlas.data.sql._
 import io.arlas.data.transform.ArlasTest
 import io.arlas.data.transform.ArlasTransformerColumns._
 import org.apache.spark.sql.types.{IntegerType, StringType}
-
 import scala.collection.immutable.ListMap
+import io.arlas.data.transform.ArlasTestHelper._
 
 class WithDurationFromIdTest extends ArlasTest {
 
   val testDF = createDataFrameWithTypes(
+    spark,
     List(
       Seq("id1", 0, 10, 100),
       Seq("id1", 90, 100, 100),
