+++ src/test/scala/io/arlas/data/transform/testdata/TestDataGenerator.scala
@@ -1,0 +1,7 @@
+package io.arlas.data.transform.testdata
+
+import org.apache.spark.sql.DataFrame
+
+trait TestDataGenerator {
+  def get(): DataFrame
+}
